+++ src/app/modules/search-parameter-group/search-parameter-group.component.ts
@@ -104,7 +104,7 @@ export class SearchParameterGroupComponent
         const match = this.resourceTypes.find((rt) => rt === value);
         if (match) {
           this.resourceType.disable({ emitEvent: false });
-          this.liveAnnoncer.announce(`Selected resource type ${value}.`);
+          this.liveAnnoncer.announce(`Selected record type ${value}.`);
         }
       });
     }
@@ -171,8 +171,8 @@ export class SearchParameterGroupComponent
   }
 
   /**
-   * Focus "Resource Type" control.
-   * This is being called from parent component when the "Add a resource type" button is clicked.
+   * Focus "Record type" control.
+   * This is being called from parent component when the "Add a record type" button is clicked.
    */
   focusResourceTypeInput(): void {
     this.resourceTypeInput.nativeElement.focus();